+++ tests/testthat/test-merge_mbox_all.R
@@ -1,0 +1,5 @@
+context("test-merge_mbox_all")
+
+test_that("multiplication works", {
+  expect_equal(2 * 2, 4)
+})